Moved the safe queue to it's own header
